Rename module to 'base-url'.
